+++ reflectable/lib/src/encoding_constants.dart
@@ -32,3 +32,7 @@ const dynamicAttribute = 1 << (flagsBit + 10);
 const classTypeAttribute = 1 << (flagsBit + 11);
 
 int kindFromEncoding(int encoding) => encoding & ((1 << flagsBit) - 1);
+
+/// Used in place of an index into a list, where a needed capability to provide
+/// the actual value is missing.
+const int NO_CAPABILITY_INDEX = -1;