+++ reflectable/test/mock_tests/check_literal_transform_test.dart
@@ -70,7 +70,7 @@ _initializeReflectable() {
         "reflectable package. Remember to set your package-root to "
         "'build/.../packages'.");
   }
-  r.data = {const prefix0.MyReflectable(): new r.ReflectorData([new r.ClassMirrorImpl(r"A", r".A", 0, const prefix0.MyReflectable(), <int>[-1], [], [], -1, {}, {}, {r"": () => new prefix0.A()}, null)], [new r.MethodMirrorImpl(r"", 64, 0, 0, [], const prefix0.MyReflectable(), null)], [], [prefix0.A], {}, {})};
+  r.data = {const prefix0.MyReflectable(): new r.ReflectorData([new r.ClassMirrorImpl(r"A", r".A", 0, const prefix0.MyReflectable(), <int>[-1], [], [], -1, {}, {}, {r"": () => new prefix0.A()}, -1, null)], [new r.MethodMirrorImpl(r"", 64, 0, 0, [], const prefix0.MyReflectable(), null)], [], [prefix0.A], {}, {}, null)};
 }
 """
   }
@@ -80,12 +80,14 @@ checkTransform(List maps) async {
   Map<String, String> inputs = maps[0];
   Map<String, String> expectedOutputs = maps[1];
   TestAggregateTransform transform = new TestAggregateTransform(inputs);
-  ReflectableTransformer transformer = new ReflectableTransformer.asPlugin(
-      new BarbackSettings(
-          {"entry_points": ["main.dart"]}, BarbackMode.RELEASE));
+  ReflectableTransformer transformer =
+      new ReflectableTransformer.asPlugin(new BarbackSettings({
+    "entry_points": ["main.dart"]
+  }, BarbackMode.RELEASE));
 
   // Test `declareOutputs`.
-  TestDeclaringTransform declaringTransform = new TestDeclaringTransform(inputs);
+  TestDeclaringTransform declaringTransform =
+      new TestDeclaringTransform(inputs);
   await transformer.declareOutputs(declaringTransform);
   expect(declaringTransform.outputs, new Set.from(expectedOutputs.keys));
   expect(declaringTransform.consumed, new Set.from([]));